forwarder/requirements.txt
awsiotsdk==1.7.1
paho-mqtt==1.6.1